rename strcmp step into comparetext
