+++ docs/img/docs/user-interface/comparetext_example_conf.jpg
