+++ docs/img/docs/user-interface/comparetext_step_form.jpg
