+++ server/tests/steps/test_comparetext.py
@@ -1,16 +1,16 @@
 import pandas as pd
 
 from tests.utils import assert_dataframes_equals
-from weaverbird.steps.strcmp import CompareTextStep
+from weaverbird.steps.comparetext import CompareTextStep
 
 
-def test_strcmp():
+def test_comparetext():
     df = pd.DataFrame(
         {'TEXT_1': ['Bbb', 'Bbb', 'Bbb', 'Bbb'], 'TEXT_2': ['bbb', 'Bbb', 'Aaa', 'cc']}
     )
 
     df_result = CompareTextStep(
-        name='strcmp',
+        name='comparetext',
         newColumnName='RESULT',
         strCol1='TEXT_1',
         strCol2='TEXT_2',
