+++ src/components/stepforms/index.ts
@@ -56,13 +56,14 @@ const StepFormsComponents: { [K in PipelineStepName]: VueConstructor<StepForm> }
   append: AppendStepForm,
   argmin: ArgminStepForm,
   argmax: ArgmaxStepForm,
-  duration: ComputeDurationStepForm,
+  comparetext: CompareTextStepForm,
   concatenate: ConcatenateStepForm,
   convert: ConvertStepForm,
   cumsum: CumSumStepForm,
   custom: CustomStepForm,
   dateextract: DateExtractStepForm,
   delete: DeleteColumnStepForm,
+  duration: ComputeDurationStepForm,
   duplicate: DuplicateColumnStepForm,
   domain: DomainStepForm,
   evolution: EvolutionStepForm,
@@ -83,7 +84,6 @@ const StepFormsComponents: { [K in PipelineStepName]: VueConstructor<StepForm> }
   split: SplitStepForm,
   sort: SortStepForm,
   statistics: StatisticsStepForm,
-  strcmp: CompareTextStepForm,
   substring: SubstringStepForm,
   todate: ToDateStepForm,
   lowercase: ToLowerStepForm,
